requirements.txt
python-telegram-bot==21.6
watchfiles==0.24.0
